pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visprobe"
version = "0.2.0"
description = "Find robustness failures in your vision models in 5 minutes"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Bilge Demirkaya"}
]
keywords = ["adversarial-robustness", "computer-vision", "deep-learning", "neural-networks", "pytorch", "testing", "fuzzing", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
    "Topic :: Security",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.19.0",
    "pillow>=8.0.0",
    "tqdm>=4.60.0",
    "pyyaml>=6.0.0",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "pytest-xdist>=2.5.0",
    "black>=23.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.10.0",
    "bandit>=1.7.0",
    "safety>=2.0.0",
    "pre-commit>=2.20.0",
]
viz = [
    "altair>=4.2.0",
]
adversarial = [
    "adversarial-robustness-toolbox>=1.18.0",
]
bayesian = [
    "scipy>=1.9.0",
    "scikit-learn>=1.0.0",
]
validation = [
    "jsonschema>=4.0.0",
]

[project.scripts]
visprobe = "visprobe.cli.cli:main"

[project.urls]
Homepage = "https://github.com/bilgedemirkaya/VisProbe"
Repository = "https://github.com/bilgedemirkaya/VisProbe"
Issues = "https://github.com/bilgedemirkaya/VisProbe/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["visprobe*"]
exclude = ["tests*"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src/visprobe",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "requires_cuda: marks tests that require CUDA",
    "requires_art: marks tests that require ART library",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/visprobe"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2
show_missing = true

# mypy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_optional = false

[[tool.mypy.overrides]]
module = "visprobe.*"
ignore_missing_imports = true

# Bandit configuration
[tool.bandit]
exclude_dirs = ["/tests", "/test_files"]
skips = ["B101", "B601"]

# Flake8 configuration (in .flake8 file, as it doesn't support pyproject.toml yet)
